refactor wrgdv3
