+++ src/scripts/modules/gooddata-writer-v3/adapters/columnsEditorDefinition.js
@@ -1,10 +1,7 @@
 import {Map, fromJS} from 'immutable';
 import {parseParameters, createConfigParameters} from '../helpers/rowParametersTable';
 import PreferencesHeader from '../react/components/PreferencesHeader';
-
-// import TitleColumnInput from '../react/components/TitleColumnInput';
-// import ColumnType from '../react/components/ColumnType';
-import ColumnSetup from '../react/components/ColumnSetup';
+import PreferencesColumn from '../react/components/PreferencesColumn';
 import {prepareColumnContext, initHeaderState} from '../helpers/ColumnDefinition';
 import ColumnDefinition from '../helpers/ColumnDefinition';
 
@@ -31,7 +28,7 @@ export default {
   columnsMappings: [
     {
       title: PreferencesHeader,
-      render: ColumnSetup
+      render: PreferencesColumn
 
     }
   ],
