+++ src/scripts/modules/gooddata-writer-v3/react/components/PreferencesColumn.jsx
